refactor: separate survey list table
